updated settings cleanup to prevent issue with the update process
Former-commit-id: 07122c535d47f3f414659013555fa826ce0e9b9c

+++ dependency-check-core/src/main/java/org/owasp/dependencycheck/utils/Settings.java
@@ -247,9 +247,11 @@ public final class Settings {
 
     /**
      * Cleans up resources to prevent memory leaks.
+     *
+     * @param deleteTemporary flag indicating whether any temporary directories generated should be removed
      */
-    public static void cleanup() {
-        if (tempDirectory != null && tempDirectory.exists()) {
+    public static void cleanup(boolean deleteTemporary) {
+        if (deleteTemporary && tempDirectory != null && tempDirectory.exists()) {
             FileUtils.delete(tempDirectory);
         }
         try {
